+++ common/src/main/java/earth/terrarium/athena/impl/platform/ModelLoaderService.java
@@ -1,12 +1,15 @@
 package earth.terrarium.athena.impl.platform;
 
 import earth.terrarium.athena.api.client.models.AthenaModelFactory;
+import earth.terrarium.athena.api.client.utils.AthenaUnbakedModelLoader;
 import net.minecraft.resources.Identifier;
+import org.jetbrains.annotations.ApiStatus;
 
+@ApiStatus.Internal
 @PlatformService
 public interface ModelLoaderService {
 
-    void register(Identifier type, AthenaModelFactory factory);
+    AthenaUnbakedModelLoader register(Identifier type, AthenaModelFactory factory);
 
     static ModelLoaderService create() {
         throw new AssertionError("Failed to create ModelLoaderService, no implementation found");