+++ fabric/src/main/java/earth/terrarium/athena/fabric/client/AthenaModelLoadingPlugin.java
@@ -1,6 +1,6 @@
-package earth.terrarium.athena.api.client.fabric;
+package earth.terrarium.athena.fabric.client;
 
-import earth.terrarium.athena.impl.platform.ModelLoaderServiceFabricImpl;
+import earth.terrarium.athena.api.client.models.FactoryManager;
 import net.fabricmc.fabric.api.client.model.loading.v1.ModelLoadingPlugin;
 import org.jspecify.annotations.NonNull;
 
@@ -10,7 +10,7 @@ public class AthenaModelLoadingPlugin implements ModelLoadingPlugin {
 
     @Override
     public void initialize(@NonNull Context context) {
-        ModelLoaderServiceFabricImpl.LOADERS.forEach((_, loader) ->
+        FactoryManager.loaders().forEach(loader ->
                 context.modifyBlockModelBeforeBake().register((model, ctx) ->
                         Objects.requireNonNullElse(loader.loadModel(ctx.state()), model)
                 )